// src/services/models/taxFarmResponse.ts
/**
 * Generated by orval v7.1.0 🍺
 * Do not edit manually.
 * EVM TAX FARM API
 * OpenAPI spec version: 1.0.0
 */

export interface TaxFarmResponse {
  readonly banner: string;
  readonly blockNumber: number;
  readonly created_at: string;
  readonly createTime: number;
  readonly creator: string;
  readonly decimals: string;
  readonly description: string;
  readonly feeTier: number;
  readonly id: number;
  readonly logo: string;
  readonly marketCap: number;
  readonly priceCurrent: string;
  readonly priceLast5m: string;
  readonly revenue: number;
  readonly status: string;
  readonly symbol: string;
  readonly telegram: string;
  readonly tokenAddress: string;
  readonly tokenName: string;
  readonly totalSupply: string;
  readonly txBurn: string;
  readonly txWithdraw: string;
  readonly uniswapPair: string;
  readonly updated_at: string;
  readonly volume5m: number;
  readonly website: string;
  readonly X: string;
}